requirements.txt
pandas>=2.2.0
numpy>=1.26.0
yfinance>=0.2.50
requests>=2.31.0
PyYAML>=6.0.0
google-analytics-data>=0.18.0
google-api-python-client>=2.160.0
google-auth>=2.38.0
